requirements.txt
librosa==0.9.2
numpy~=1.21
opencv-python==4.7.0.72
ray==2.2.0
scipy==1.9.3
soundfile==0.11.0
tensorboard
torch==1.13.1
tqdm~=4.64.1
pytorch_msssim==0.2.1
torchvision==0.14.1
moviepy==1.0.3
pandas~=1.5.2
imageio~=2.25.0
pillow~=9.3.0
matplotlib~=3.6.3
parameterized~=0.8.1